feat: Plat-6165 implement transfertransaction rpc method (#63)
* feat: plat-6154 Add support for Solana local signer

* chore: code improvements

* chore: update cargo lock

* feat: plat-6158 implement solana provider

* chore: code improvements

* chore: add logic for initializing relayer and running checks

* feat: plat-6159 extend relayer solana policies

* chore: code improvements|

* feat: add support for solana policies and validate them upon service start

* chore: partial

* chore: improvements

* chore: remove cargo feature

* chore: update cargo

* feat: plat-6167 implement signTransaction rpc method

* chore: add tests

* chore: impr

* chore: improvements

* chore: add validate token transfers logic

* chore: lint

* chore: refactor max fee

* chore: add tests

* chore: add transfer amout validation

* chore: Add fee validation

* chore: improvements

* chore: add tests

* chore: format and cargo update

* chore: update docs

* feat: plat-6164 implement fee_eastimate rpc method

* chore: impr

* chore: revert overwrite

* chore: add test script

* chore: improvements

* chore: updates

* chore: impr

* chore: add comments to ignore security findings

* feat: plat-6165 implement transfer transaction method

* chore: improvements

* chore: impl jupiter mock service

* chore: add prepare transaction impl

* chore: add tests for prepare transaction

* chore: add rpc methods rust docs

* chore: add balance checks

* chore: refactor

* chore: add more tests

* chore: lint

* chore: add more util tests

* chore: clippy

* chore: add noboost comments

* chore: fix build

+++ src/models/rpc/solana/mod.rs
@@ -4,6 +4,7 @@ use solana_sdk::transaction::Transaction;
 use thiserror::Error;
 
 #[derive(Debug, Error, Deserialize, Serialize)]
+#[allow(clippy::enum_variant_names)]
 pub enum SolanaEncodingError {
     #[error("Failed to serialize transaction: {0}")]
     Serialization(String),
@@ -68,7 +69,7 @@ pub struct FeeEstimateResult {
 // transferTransaction
 #[derive(Debug, Deserialize, Serialize, PartialEq)]
 pub struct TransferTransactionRequestParams {
-    pub amount: usize,
+    pub amount: u64,
     pub token: String,
     pub source: String,
     pub destination: String,
